customer-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
supabase==2.28.0
python-dotenv==1.0.0
httpx>=0.26,<0.29
pydantic==2.12.5
